+++ flutter_inappwebview/example/test/utils/new_api_coverage_test.dart
@@ -52,6 +52,7 @@ void main() {
           InAppWebViewSettingsProperty.supportsAdaptiveImageGlyph,
           InAppWebViewSettingsProperty.writingToolsBehavior,
           InAppWebViewSettingsProperty.showsSystemScreenTimeBlockingView,
+          InAppWebViewSettingsProperty.useOnInsertInputSuggestion,
         ]),
       );
     });