+++ flutter_inappwebview_platform_interface/lib/src/in_app_webview/platform_headless_in_app_webview.dart
@@ -85,6 +85,7 @@ class PlatformHeadlessInAppWebViewCreationParams
     super.onMicrophoneCaptureStateChanged,
     super.onContentSizeChanged,
     super.onShowFileChooser,
+    super.onInsertInputSuggestion,
     super.initialUrlRequest,
     super.initialFile,
     super.initialData,