+++ flutter_inappwebview_platform_interface/lib/src/in_app_webview/platform_inappwebview_widget.dart
@@ -93,6 +93,7 @@ class PlatformInAppWebViewWidgetCreationParams
     super.onMicrophoneCaptureStateChanged,
     super.onContentSizeChanged,
     super.onShowFileChooser,
+    super.onInsertInputSuggestion,
     super.initialUrlRequest,
     super.initialFile,
     super.initialData,